bus/hp_dio/98543: convert to anonymous namespace (nw)
